requirements.txt
pytest==7.2.0
redis=4.3.4
requests==2.28.1
selenium==4.6.0
Appium-Python-Client==2.7.1
Appium-UIAutomation==0.1.0
PyMySQL==1.0.2
allure-pytest==2.11.1
allure-python-commons==2.11.1
PyYAML==6.0